renamed and moved CDMemberMapper
